+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.13.0
 
 * show search aggregations, fixes [#247](https://github.com/cars10/elasticvue/issues/247)
+* can filter indices by alias, fixes [#337](https://github.com/cars10/elasticvue/issues/337)
 * i18n: adds traditional chinese, thanks @PeterDaveHello
 * i18n: adds korean, thanks @webee0317